Visualiserer med riktig id

+++ src/utils/callbacks.py
@@ -26,9 +26,10 @@ def get_model_checkpoint(config):
 
 
 class LogPredictionsCallback(Callback):
-    def __init__(self, log_every_n_epochs=5, artifact_dir="val_predictions"):
+    def __init__(self, log_every_n_epochs=5, artifact_dir="val_predictions", extra_artifacts=None):
         self.log_every_n_epochs = log_every_n_epochs
         self.artifact_dir = artifact_dir
+        self.extra_artifacts = extra_artifacts or []
 
     def on_validation_epoch_end(self, trainer, pl_module):
         if (trainer.current_epoch % self.log_every_n_epochs) != 0:
@@ -36,7 +37,7 @@ class LogPredictionsCallback(Callback):
 
         dataloader = trainer.datamodule.val_dataloader()
         batch = next(iter(dataloader))
-        x, y = batch
+        x, y, filenames = batch
         x, y = x.to(pl_module.device), y.to(pl_module.device)
         logits = pl_module(x)
         preds = torch.sigmoid(logits) > 0.5
@@ -45,33 +46,39 @@ class LogPredictionsCallback(Callback):
         max_images = min(num_images, x.shape[0])
 
         for i in range(max_images):
-            image = x[i, :3].detach().cpu()    # RGB
-            dom = x[i, 3].detach().cpu()       # DOM-kanal (må bruke indeks [0, 3])
-            target = y[i].detach().cpu()
-            pred = preds[i].detach().cpu()
+            filename = filenames[i]
+            display_always = filename in self.extra_artifacts
+            display_default = i < num_images
 
-            fig, axs = plt.subplots(1, 4, figsize=(14, 4))
-            axs[0].imshow(image.permute(1, 2, 0))
-            axs[0].set_title("Input RGB")
-            axs[1].imshow(dom, cmap="gray")
-            axs[1].set_title("Input DOM")
-            axs[2].imshow(target.squeeze(), cmap="gray")
-            axs[2].set_title("Target mask")
-            axs[3].imshow(pred.squeeze(), cmap="gray")
-            axs[3].set_title("Predicted mask")
-            for ax in axs:
-                ax.axis("off")
-            fig.tight_layout()
+            if display_default or display_always:
+                
+                image = x[i, :3].detach().cpu()    # RGB
+                dom = x[i, 3].detach().cpu()       # DOM-kanal (må bruke indeks [0, 3])
+                target = y[i].detach().cpu()
+                pred = preds[i].detach().cpu()
 
-            image_dir = os.path.join(self.artifact_dir, f"image_{i}")
-            os.makedirs(image_dir, exist_ok=True)
+                fig, axs = plt.subplots(1, 4, figsize=(14, 4))
+                axs[0].imshow(image.permute(1, 2, 0))
+                axs[0].set_title("Input RGB")
+                axs[1].imshow(dom, cmap="gray")
+                axs[1].set_title("Input DOM")
+                axs[2].imshow(target.squeeze(), cmap="gray")
+                axs[2].set_title("Target mask")
+                axs[3].imshow(pred.squeeze(), cmap="gray")
+                axs[3].set_title("Predicted mask")
+                for ax in axs:
+                    ax.axis("off")
+                fig.tight_layout()
 
-            fname = os.path.join(image_dir, f"epoch_{trainer.current_epoch}.png")
-            fig.savefig(fname)
-            plt.close(fig)
+                image_id = os.path.splitext(os.path.basename(filename))[0]
+                image_dir = os.path.join(self.artifact_dir, image_id)
+                os.makedirs(image_dir, exist_ok=True)
 
-            mlflow_client = trainer.logger.experiment
-            run_id = trainer.logger.run_id
-            mlflow_client.log_artifact(run_id, fname, artifact_path=image_dir)
+                fname = os.path.join(image_dir, f"epoch_{trainer.current_epoch}.png")
+                fig.savefig(fname)
+                plt.close(fig)
 
-            os.remove(fname)
+                mlflow_client = trainer.logger.experiment
+                run_id = trainer.logger.run_id
+                mlflow_client.log_artifact(run_id, fname, artifact_path=image_dir)
+                os.remove(fname)
